Fix important dune warnings
Mainly added cases to pattern matching so it will print a bug message in
the case that it cannot match.

There are still a lot of warnings around unused functions and types. I
don't want to remove these blindly - they could be useful in places that
I can't recognize at the moment.