Add new variables to the sequent on assertion
When using the `assert' tactic, if a new variable name is specified in
the context of an atomic formula then we replace it with a fresh nominal
and add it to the sequent. It will not add it to the restricted set of
the context variable - as that is expected to be asserted via an
abstraction.