Fix schema instantiation checking
This fixes the previous method by which Adelfa determined if an
existentially quantified context variable could be insantiated into a
context expression which is a repetition of the schema.

As part of the change, unit tests for the new process are added as well.

Specifically, the previous instance check would always succeed if the
existentially quantified context variable was quantified by the same
schema as the universally quantified one. Further, when checking if an
explicit portion of the context is an instance of a block, we try to
unify the explicit portion with the schema's definition. In preparation
for the unification process, nominal variables are substituted for the
schematic variables in the definition. This substitution did not
properly replace all instances of the schematic variables.